+++ CHANGES.md
@@ -7,6 +7,10 @@
   https://github.com/joblib/threadpoolctl/pull/48
   https://github.com/joblib/threadpoolctl/pull/60
 
+- When threadpoolctl finds libomp (LLVM OpenMP) and libiomp (Intel OpenMP)
+  both loaded, a warning is raised to recall that using threadpoolctl with
+  this mix of OpenMP libraries may cause crashes or deadlocks.
+  https://github.com/joblib/threadpoolctl/pull/49
 
 1.1.0 (2019-09-12)
 ==================